backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
elasticsearch==8.11.0
PyJWT==2.8.0
bcrypt==4.1.2
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
python-multipart==0.0.6
Pillow==10.1.0
boto3==1.34.0
django-storages==1.14.2
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
